move states to seperated packages
