+++ app/src/main/kotlin/com/wire/android/ui/home/messagecomposer/MessageComposer.kt
@@ -2,7 +2,6 @@ package com.wire.android.ui.home.messagecomposer
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.animateContentSize
 import androidx.compose.animation.core.animateFloat
 import androidx.compose.animation.core.updateTransition
@@ -11,7 +10,6 @@ import androidx.compose.animation.fadeOut
 import androidx.compose.animation.slideOutVertically
 import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.detectTapGestures
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
@@ -23,15 +21,11 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.Divider
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.Surface
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -41,15 +35,9 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.focus.FocusManager
-import androidx.compose.ui.focus.onFocusChanged
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.Dp
@@ -57,22 +45,16 @@ import androidx.compose.ui.unit.dp
 import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.constraintlayout.compose.Dimension
 import com.wire.android.R
-import com.wire.android.ui.common.button.WireButtonState
-import com.wire.android.ui.common.button.WireIconButton
 import com.wire.android.ui.common.dimensions
-import com.wire.android.ui.common.textfield.WireTextField
-import com.wire.android.ui.common.textfield.wireTextFieldColors
 import com.wire.android.ui.home.conversations.model.AttachmentBundle
 import com.wire.android.ui.home.messagecomposer.attachment.AttachmentOptionsComponent
-import com.wire.android.ui.home.messagecomposer.button.AddEmojiAction
-import com.wire.android.ui.home.messagecomposer.button.AddGifAction
-import com.wire.android.ui.home.messagecomposer.button.AddMentionAction
 import com.wire.android.ui.home.messagecomposer.button.AdditionalOptionButton
-import com.wire.android.ui.home.messagecomposer.button.RichTextEditingAction
-import com.wire.android.ui.home.messagecomposer.button.TakePictureAction
+import com.wire.android.ui.home.messagecomposer.button.CollapseIconButton
+import com.wire.android.ui.home.messagecomposer.button.ScheduleMessageButton
+import com.wire.android.ui.home.messagecomposer.button.SendButton
+import com.wire.android.ui.home.messagecomposer.input.MessageComposerInput
 import com.wire.android.ui.theme.wireColorScheme
 import com.wire.android.ui.theme.wireDimensions
-import com.wire.android.ui.theme.wireTypography
 
 private val DEFAULT_KEYBOARD_TOP_SCREEN_OFFSET = 250.dp
 
@@ -130,7 +112,7 @@ fun MessageComposer(
 private fun MessageComposer(
     content: @Composable () -> Unit,
     screenHeight: Dp,
-    messageComposerState: MessageComposerInnerState,
+    messageComposerState: MessageComposerState,
     messageText: TextFieldValue,
     onMessageChanged: (TextFieldValue) -> Unit,
     onSendButtonClicked: () -> Unit,
@@ -374,7 +356,7 @@ private fun MessageComposer(
                 ) {
                     Divider()
                     AttachmentOptionsComponent(
-                        messageComposerState.attachmentInnerState,
+                        messageComposerState.attachmentState,
                         onSendAttachment,
                         onError,
                         Modifier.align(Alignment.Center)
@@ -382,122 +364,6 @@ private fun MessageComposer(
                 }
             }
         }
-    }
-}
-
-//if attachment is visible we want to align the bottom of the compose actions
-// to top of the guide line
-@Composable
-private fun CollapseIconButton(onCollapseClick: () -> Unit, modifier: Modifier = Modifier, collapseRotation: Float = 0f) {
-    IconButton(
-        onClick = onCollapseClick,
-        modifier = modifier.size(20.dp)
-    ) {
-        Icon(
-            painter = painterResource(id = R.drawable.ic_collapse),
-            contentDescription = stringResource(R.string.content_description_drop_down_icon),
-            tint = MaterialTheme.wireColorScheme.onSecondaryButtonDisabled,
-            modifier = Modifier.rotate(collapseRotation)
-        )
-    }
-}
-
-@Composable
-private fun ScheduleMessageButton() {
-    IconButton(onClick = { }) {
-        Icon(
-            painter = painterResource(id = R.drawable.ic_timer),
-            contentDescription = stringResource(R.string.content_description_back_button),
-        )
-    }
-}
-
-@Composable
-private fun MessageComposerInput(
-    messageText: TextFieldValue,
-    onMessageTextChanged: (TextFieldValue) -> Unit,
-    messageComposerInputState: MessageComposeInputState,
-    onIsFocused: () -> Unit,
-    onNotFocused: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    WireTextField(
-        value = messageText,
-        onValueChange = onMessageTextChanged,
-        colors = wireTextFieldColors(
-            borderColor = Color.Transparent,
-            focusColor = Color.Transparent
-        ),
-        singleLine = messageComposerInputState == MessageComposeInputState.Enabled,
-        maxLines = Int.MAX_VALUE,
-        textStyle = MaterialTheme.wireTypography.body01,
-        // Add a extra space so that the a cursor is placed one space before "Type a message"
-        placeholderText = " " + stringResource(R.string.label_type_a_message),
-        modifier = modifier.then(
-            Modifier.onFocusChanged { focusState ->
-                if (focusState.isFocused) {
-                    onIsFocused()
-                } else {
-                    onNotFocused()
-                }
-            }
-        )
-    )
-}
-
-@Composable
-private fun SendButton(
-    isEnabled: Boolean,
-    onSendButtonClicked: () -> Unit
-) {
-    IconButton(
-        onClick = { if (isEnabled) onSendButtonClicked() },
-        enabled = isEnabled
-    ) {
-        Box(
-            contentAlignment = Alignment.Center,
-            modifier = Modifier
-                .size(40.dp)
-                .clip(CircleShape)
-                .background(
-                    animateColorAsState(
-                        when {
-                            isEnabled -> MaterialTheme.colorScheme.primary
-                            else -> MaterialTheme.wireColorScheme.onSecondaryButtonDisabled
-                        }
-                    ).value
-                )
-        ) {
-            Icon(
-                painter = painterResource(id = R.drawable.ic_send),
-                contentDescription = stringResource(R.string.content_description_back_button),
-                tint = MaterialTheme.wireColorScheme.surface
-            )
-        }
-    }
-}
-
-@Composable
-private fun MessageComposeActions(
-    messageComposerState: MessageComposerInnerState,
-    focusManager: FocusManager
-) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceEvenly,
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(56.dp)
-    ) {
-        AdditionalOptionButton(messageComposerState.attachmentOptionsDisplayed) {
-            focusManager.clearFocus()
-            messageComposerState.toggleAttachmentOptionsVisibility()
-        }
-        RichTextEditingAction()
-        AddEmojiAction()
-        AddGifAction()
-        AddMentionAction()
-        TakePictureAction()
     }
 }
