+++ app/src/main/kotlin/com/wire/android/ui/home/messagecomposer/button/SendButton.kt
@@ -1,4 +1,51 @@
 package com.wire.android.ui.home.messagecomposer.button
 
-class SendButton {
+import androidx.compose.animation.animateColorAsState
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.wire.android.R
+import com.wire.android.ui.theme.wireColorScheme
+
+@Composable
+fun SendButton(
+    isEnabled: Boolean,
+    onSendButtonClicked: () -> Unit
+) {
+    IconButton(
+        onClick = { if (isEnabled) onSendButtonClicked() },
+        enabled = isEnabled
+    ) {
+        Box(
+            contentAlignment = Alignment.Center,
+            modifier = Modifier
+                .size(40.dp)
+                .clip(CircleShape)
+                .background(
+                    animateColorAsState(
+                        when {
+                            isEnabled -> MaterialTheme.colorScheme.primary
+                            else -> MaterialTheme.wireColorScheme.onSecondaryButtonDisabled
+                        }
+                    ).value
+                )
+        ) {
+            Icon(
+                painter = painterResource(id = R.drawable.ic_send),
+                contentDescription = stringResource(R.string.content_description_back_button),
+                tint = MaterialTheme.wireColorScheme.surface
+            )
+        }
+    }
 }
